requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
docopt==0.6.2
Flask
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.27.8
packaging==23.1
Pillow==9.5.0
pipreqs==0.4.13
psycopg2==2.9.6
pytesseract==0.3.10
python-dotenv
requests==2.31.0
tqdm==4.65.0
urllib3==2.0.3
Werkzeug
yarg==0.1.9
yarl==1.9.2

